Fix sorting order of table
