+++ src/calculate.cpp
@@ -217,3 +217,20 @@ std::string Calculate::favorite_type(Storage storage) {
 
     return favorite_type;
 }
+
+bool Calculate::compare_date(const Drink &a, const Drink &b) {
+    /*
+     * Determine if second date is greater than the first date.
+     * @return: True if second date is more recent than the first date. Else, false.
+     */
+
+    if (a.drink_year < b.drink_year) {
+        return true;
+    } else if (a.drink_year == b.drink_year && a.drink_month < b.drink_month) {
+        return true;
+    } else if (a.drink_year == b.drink_year && a.drink_month == b.drink_month && a.drink_day < b.drink_day) {
+        return true;
+    } else {
+        return false;
+    }
+}
