+++ src/mainwindow.cpp
@@ -147,6 +147,7 @@ MainWindow::MainWindow(QWidget *parent)
     ui->drinkLogTable->setColumnWidth(8, 55);
     ui->drinkLogTable->setColumnHidden(9, true);  // Hide ID column
     ui->drinkLogTable->setColumnHidden(10, true);  // Hide Timestamp column
+    ui->drinkLogTable->setColumnHidden(11, true);  // Hide Sort column
     QHeaderView* drink_log_header = ui->drinkLogTable->horizontalHeader();
     drink_log_header->setSectionResizeMode(1, QHeaderView::Stretch);
     drink_log_header->setSectionResizeMode(2, QHeaderView::Stretch);
@@ -322,36 +323,39 @@ void MainWindow::update_table() {
     std::string filter_category = ui->filterCategoryInput->currentText().toStdString();
     std::string filter_text = ui->filterTextInput->currentText().toStdString();
 
-    std::vector<Drink> beers = Database::filter(filter_category, filter_text, storage);
+    std::vector<Drink> drinks = Database::filter(filter_category, filter_text, storage);
+    drinks = Database::sort_by_date_id(drinks);
 
     ui->drinkLogTable->setRowCount(0);
-    for (const auto& beer : beers) {
+    for (const auto& drink : drinks) {
         int table_row_num = ui->drinkLogTable->rowCount();
         ui->drinkLogTable->insertRow(table_row_num);
-        QDate date = QDate(beer.drink_year, beer.drink_month, beer.drink_day);
+        QDate date = QDate(drink.drink_year, drink.drink_month, drink.drink_day);
         auto *date_qtw = new QTableWidgetItem;
-        auto *name = new QTableWidgetItem(beer.name.c_str());
-        auto *type = new QTableWidgetItem(beer.type.c_str());
-        auto *subtype = new QTableWidgetItem(beer.subtype.c_str());
-        auto *producer = new QTableWidgetItem(beer.producer.c_str());
-        auto *abv = new QTableWidgetItem(double_to_string(beer.abv).c_str());
-        auto *size = new QTableWidgetItem(double_to_string(beer.size).c_str());
-        auto *rating = new QTableWidgetItem(std::to_string(beer.rating).c_str());
+        auto *name = new QTableWidgetItem(drink.name.c_str());
+        auto *type = new QTableWidgetItem(drink.type.c_str());
+        auto *subtype = new QTableWidgetItem(drink.subtype.c_str());
+        auto *producer = new QTableWidgetItem(drink.producer.c_str());
+        auto *abv = new QTableWidgetItem(double_to_string(drink.abv).c_str());
+        auto *size = new QTableWidgetItem(double_to_string(drink.size).c_str());
+        auto *rating = new QTableWidgetItem(std::to_string(drink.rating).c_str());
         auto *id = new QTableWidgetItem;
-        auto *timestamp = new QTableWidgetItem(beer.timestamp.c_str());
+        auto *timestamp = new QTableWidgetItem(drink.timestamp.c_str());
+        auto *sort_order =  new QTableWidgetItem(drink.sort_order);
 
         // Sort ID numerically
-        id->setData(Qt::DisplayRole, beer.id);
+        id->setData(Qt::DisplayRole, drink.id);
+        sort_order->setData(Qt::DisplayRole, drink.sort_order);
         date_qtw->setData(Qt::DisplayRole, date);
 
-        std::string notes = beer.notes;
+        std::string notes = drink.notes;
 
         // Handle blank IBU
         auto *ibu = new QTableWidgetItem;
-        if (beer.ibu == -1.0) {  // -1 denotes no IBU value
+        if (drink.ibu == -1.0) {  // -1 denotes no IBU value
             *ibu = QTableWidgetItem("");
         } else {
-            *ibu = QTableWidgetItem(double_to_string(beer.ibu).c_str());
+            *ibu = QTableWidgetItem(double_to_string(drink.ibu).c_str());
         }
 
         ui->drinkLogTable->setItem(table_row_num, 0, date_qtw);
@@ -365,6 +369,7 @@ void MainWindow::update_table() {
         ui->drinkLogTable->setItem(table_row_num, 8, rating);
         ui->drinkLogTable->setItem(table_row_num, 9, id);
         ui->drinkLogTable->setItem(table_row_num, 10, timestamp);
+        ui->drinkLogTable->setItem(table_row_num, 11, sort_order);
     }
     reset_table_sort();
 }
@@ -496,239 +501,6 @@ void MainWindow::delete_row() {
     update_beer_fields();
     update_liquor_fields();
     update_wine_fields();
-}
-
-std::vector<std::set<QString>> MainWindow::generate_filter_item_sets() {
-    /*
-     * Generate sets containing unique filter values.
-     * @return filter_values: A vector of sets containing unique values to pass to the filter dropdown.
-     */
-
-    std::vector<std::set<QString>> filter_values;
-
-    std::set<QString> drink_names;
-    std::set<QString> drink_types;
-    std::set<QString> drink_subtypes;
-    std::set<QString> producers;
-    std::set<QString> ratings;
-
-    ui->filterTextInput->clear();
-
-    // This fixes crashes when changing filters with rows selected.
-    QSignalBlocker filterTextInputSignalBlocker(ui->filterTextInput);
-
-    // Add items to the sets
-    std::vector<Drink> all_beers = Database::read(Database::path(), storage);
-    for (const auto& beer : all_beers) {
-        QString drink_name = QString::fromStdString(beer.name);
-        QString drink_type = QString::fromStdString(beer.type);
-        QString drink_subtype = QString::fromStdString(beer.subtype);
-        QString producer = QString::fromStdString(beer.producer);
-        QString rating = QString::fromStdString(std::to_string(beer.rating));
-
-        drink_names.insert(drink_name);
-        drink_types.insert(drink_type);
-        drink_subtypes.insert(drink_subtype);
-        producers.insert(producer);
-        ratings.insert(rating);
-    }
-    filter_values.push_back(drink_names);
-    filter_values.push_back(drink_types);
-    filter_values.push_back(drink_subtypes);
-    filter_values.push_back(producers);
-    filter_values.push_back(ratings);
-
-    return filter_values;
-}
-
-void MainWindow::populate_filter_menus(const std::string& filter_type) {
-    /*
-     * Populate the filter menus depending on user selection.
-     * @param filter type: Type of filter to use. Options in beer name (name), beer type (type) and brewery.
-     */
-
-    std::vector<std::set<QString>> filter_values = generate_filter_item_sets();
-
-    if (filter_type == "Name") {
-        for (const auto& name : filter_values.at(0)) {
-            ui->filterTextInput->addItem(name);
-        }
-    } else if (filter_type == "Type") {
-        for (const auto& type : filter_values.at(1)) {
-            ui->filterTextInput->addItem(type);
-        }
-    } else if (filter_type == "Subtype") {
-        for (const auto& subtype : filter_values.at(2)) {
-            if (!subtype.isEmpty()) {
-                ui->filterTextInput->addItem(subtype);
-            }
-        }
-    } else if (filter_type == "Producer") {
-        for (const auto& brewery : filter_values.at(3)) {
-            ui->filterTextInput->addItem(brewery);
-        }
-    } else if (filter_type == "Rating") {
-        for (const auto& rating : filter_values.at(4)) {
-            ui->filterTextInput->addItem(rating);
-        }
-    }
-}
-
-void MainWindow::enable_filter_text(const QString&) {
-    /*
-     * Enable the filter text combobox when user selects anything other than None in the filter category box.
-     */
-
-    if (ui->filterCategoryInput->currentText().toStdString() == "None")
-        ui->filterTextInput->setDisabled(true);
-    else
-        ui->filterTextInput->setEnabled(true);
-
-    std::string filter_type = ui->filterCategoryInput->currentText().toStdString();
-    populate_filter_menus(filter_type);
-    update_table();
-}
-
-void MainWindow::changed_filter_text(const QString &) {
-    /*
-     * Update the table when user selects a value from the filterText combobox.
-     */
-
-    update_table();
-}
-
-void MainWindow::update_beer_fields() {
-    /*
-     * Read rows in the DB and populate the brewery, type, and name dropdowns with unique values.
-     */
-
-    std::set<QString> breweries;
-    std::set<QString> types;
-    std::set<QString> names;
-    std::vector<Drink> all_beers = Database::filter("Alcohol Type", "Beer", storage);
-
-    // Block signals to avoid crashing
-    QSignalBlocker brewery_signal_blocker(ui->beerBreweryInput);
-    QSignalBlocker type_signal_blocker(ui->beerTypeInput);
-    QSignalBlocker name_signal_blocker(ui->beerNameInput);
-
-    ui->beerBreweryInput->clear();
-    ui->beerTypeInput->clear();
-    ui->beerNameInput->clear();
-
-    for (const auto& beer : all_beers) {
-        QString brewery_name = QString::fromStdString(beer.producer);
-        QString beer_type = QString::fromStdString(beer.type);
-        QString beer_name = QString::fromStdString(beer.name);
-
-        breweries.insert(brewery_name);
-        types.insert(beer_type);
-        names.insert(beer_name);
-    }
-
-    for (const auto& brewery : breweries) {
-        ui->beerBreweryInput->addItem(brewery);
-    }
-
-    for (const auto& type : types) {
-        ui->beerTypeInput->addItem(type);
-    }
-
-    for (const auto& name : names) {
-        ui->beerNameInput->addItem(name);
-    }
-}
-
-void MainWindow::update_liquor_fields() {
-    /*
-     * Read rows in the DB and populate the distiller, type, and name dropdowns with unique values.
-     */
-
-    std::set<QString> distilleries;
-    std::set<QString> types;
-    std::set<QString> names;
-
-    std::vector<Drink> all_liquor = Database::filter("Alcohol Type", "Liquor", storage);
-
-    // Block signals to avoid crashing
-    QSignalBlocker brewery_signal_blocker(ui->liquorDistillerInput);
-    QSignalBlocker type_signal_blocker(ui->liquorTypeInput);
-    QSignalBlocker name_signal_blocker(ui->liquorNameInput);
-
-    ui->liquorDistillerInput->clear();
-    ui->liquorDistillerInput->setCurrentText("");
-    ui->liquorTypeInput->clear();
-    ui->liquorTypeInput->setCurrentText("");
-    ui->liquorNameInput->clear();
-    ui->liquorNameInput->setCurrentText("");
-
-    for (const auto& liquor : all_liquor) {
-        QString distiller_name = QString::fromStdString(liquor.producer);
-        QString liquor_type = QString::fromStdString(liquor.type);
-        QString liquor_name = QString::fromStdString(liquor.name);
-
-        distilleries.insert(distiller_name);
-        types.insert(liquor_type);
-        names.insert(liquor_name);
-    }
-
-    for (const auto& distillery : distilleries) {
-        ui->liquorDistillerInput->addItem(distillery);
-    }
-
-    for (const auto& type : types) {
-        ui->liquorTypeInput->addItem(type);
-    }
-
-    for (const auto& name : names) {
-        ui->liquorNameInput->addItem(name);
-    }
-}
-
-void MainWindow::update_wine_fields() {
-    /*
-     * Read rows in the DB and populate the winery, type, and name dropdowns with unique values.
-     */
-
-    std::set<QString> wineries;
-    std::set<QString> types;
-    std::set<QString> names;
-
-    std::vector<Drink> all_wine = Database::filter("Alcohol Type", "Wine", storage);
-
-    // Block signals to avoid crashing
-    QSignalBlocker winery_signal_blocker(ui->wineryInput);
-    QSignalBlocker type_signal_blocker(ui->wineTypeInput);
-    QSignalBlocker name_signal_blocker(ui->wineNameInput);
-
-    ui->wineryInput->clear();
-    ui->wineryInput->setCurrentText("");
-    ui->wineTypeInput->clear();
-    ui->wineTypeInput->setCurrentText("");
-    ui->wineNameInput->clear();
-    ui->wineNameInput->setCurrentText("");
-
-    for (const auto& wine : all_wine) {
-        QString winery_name = QString::fromStdString(wine.producer);
-        QString wine_type = QString::fromStdString(wine.type);
-        QString wine_name = QString::fromStdString(wine.name);
-
-        wineries.insert(winery_name);
-        types.insert(wine_type);
-        names.insert(wine_name);
-    }
-
-    for (const auto& winery : wineries) {
-        ui->wineryInput->addItem(winery);
-    }
-
-    for (const auto& type : types) {
-        ui->wineTypeInput->addItem(type);
-    }
-
-    for (const auto& name : names) {
-        ui->wineNameInput->addItem(name);
-    }
 }
 
 void MainWindow::open_about_dialog() {
@@ -907,7 +679,7 @@ void MainWindow::reset_table_sort() {
      * Reset table sort to default, by datetime descending.
      */
 
-    ui->drinkLogTable->sortItems(0, Qt::DescendingOrder);
+    ui->drinkLogTable->sortItems(11, Qt::DescendingOrder);
 }
 
 double MainWindow::update_oz_alcohol_consumed_this_week(const std::vector<Drink>& beers_this_week) {
@@ -1001,113 +773,6 @@ void MainWindow::name_input_changed(const QString&) {
     name_input_changed();
 }
 
-void MainWindow::update_beer_names_producers() {
-    /*
-     * Update the names and breweries fields of the beer tab.
-     */
-
-    std::set<QString> drink_names;
-    std::set<QString> producer_names;
-    std::string input_type = ui->beerTypeInput->currentText().toStdString();
-    std::cout << input_type << std::endl;
-    std::vector<Drink> selected_beers = Database::get_beers_by_type(storage, input_type);
-
-    // This fixes crashes when changing with rows selected.
-    QSignalBlocker name_input_signal_blocker(ui->beerNameInput);
-    QSignalBlocker brewery_input_signal_blocker(ui->beerBreweryInput);
-
-    ui->beerAbvInput->setValue(0.0);
-    ui->beerIbuInput->setValue(0.0);
-    ui->beerSizeInput->setValue(0);
-    ui->beerRatingInput->setValue(0);
-    ui->beerNameInput->clear();
-    ui->beerBreweryInput->clear();
-
-    for (const auto& selected_beer : selected_beers) {
-        drink_names.insert(QString::fromStdString(selected_beer.name));
-        producer_names.insert(QString::fromStdString(selected_beer.producer));
-    }
-
-    for (const auto& brewery : producer_names) {
-        ui->beerBreweryInput->addItem(brewery);
-    }
-
-    for (const auto& beer : drink_names) {
-        ui->beerNameInput->addItem(beer);
-    }
-}
-
-void MainWindow::update_liquor_names_producers() {
-    /*
-     * Update the names and distillery fields of the liquor tab.
-     */
-
-    std::string alcohol_type = get_current_tab();
-    std::set<QString> drink_names;
-    std::set<QString> producer_names;
-
-    std::string input_type = ui->liquorTypeInput->currentText().toStdString();
-    std::vector<Drink> selected_drinks = Database::get_beers_by_type(storage, input_type);
-
-    // This fixes crashes when changing with rows selected.
-    QSignalBlocker name_input_signal_blocker(ui->liquorNameInput);
-    QSignalBlocker brewery_input_signal_blocker(ui->liquorDistillerInput);
-
-    ui->liquorAbvInput->setValue(0.0);
-    ui->liquorSizeInput->setValue(0);
-    ui->liquorRatingInput->setValue(0);
-    ui->liquorNameInput->clear();
-    ui->liquorDistillerInput->clear();
-
-    for (const auto& selected_drink : selected_drinks) {
-        drink_names.insert(QString::fromStdString(selected_drink.name));
-        producer_names.insert(QString::fromStdString(selected_drink.producer));
-    }
-
-    for (const auto& distillery : producer_names) {
-        ui->liquorDistillerInput->addItem(distillery);
-    }
-
-    for (const auto& drink_name : drink_names) {
-        ui->liquorNameInput->addItem(drink_name);
-    }
-}
-
-void MainWindow::update_wine_names_producers() {
-    /*
-     * Update the names and winery fields of the wine tab.
-     */
-
-    std::set<QString> drink_names;
-    std::set<QString> producer_names;
-
-    std::string input_type = ui->wineTypeInput->currentText().toStdString();
-    std::vector<Drink> selected_drinks = Database::get_beers_by_type(storage, input_type);
-
-    // This fixes crashes when changing with rows selected.
-    QSignalBlocker name_input_signal_blocker(ui->wineNameInput);
-    QSignalBlocker winery_input_signal_blocker(ui->wineryInput);
-
-    ui->wineAbvInput->setValue(0.0);
-    ui->wineSizeInput->setValue(0);
-    ui->wineRatingInput->setValue(0);
-    ui->wineNameInput->clear();
-    ui->wineryInput->clear();
-
-    for (const auto& selected_drink : selected_drinks) {
-        drink_names.insert(QString::fromStdString(selected_drink.name));
-        producer_names.insert(QString::fromStdString(selected_drink.producer));
-    }
-
-    for (const auto& winery : producer_names) {
-        ui->wineryInput->addItem(winery);
-    }
-
-    for (const auto& drink_name : drink_names) {
-        ui->wineNameInput->addItem(drink_name);
-    }
-}
-
 void MainWindow::type_input_changed(const QString &) {
     /*
      * Change the beer attributes based on the type of beer selected in the typeInput field.
@@ -1127,109 +792,6 @@ void MainWindow::type_input_changed(const QString &) {
     name_input_changed();
 }
 
-void MainWindow::update_beer_names_types() {
-    /*
-     * Update the name and type on the beer tab.
-     */
-
-    std::string input_brewery = ui->beerBreweryInput->currentText().toStdString();
-    std::vector<Drink> selected_beers = Database::get_beers_by_brewery(storage, input_brewery);
-    std::set<QString> beer_names;
-    std::set<QString> types;
-
-    // This fixes crashes when changing with rows selected.
-    QSignalBlocker name_input_signal_blocker(ui->beerNameInput);
-    QSignalBlocker type_input_signal_blocker(ui->beerTypeInput);
-
-    ui->beerAbvInput->setValue(0.0);
-    ui->beerIbuInput->setValue(0.0);
-    ui->beerSizeInput->setValue(0);
-    ui->beerRatingInput->setValue(0);
-    ui->beerNameInput->clear();
-    ui->beerTypeInput->clear();
-
-    for (const auto& selected_beer : selected_beers) {
-        beer_names.insert(QString::fromStdString(selected_beer.name));
-        types.insert(QString::fromStdString(selected_beer.type));
-    }
-
-    for (const auto& name : beer_names) {
-        ui->beerNameInput->addItem(name);
-    }
-
-    for (const auto& beer_type : types) {
-        ui->beerTypeInput->addItem(beer_type);
-    }
-}
-
-void MainWindow::update_liquor_names_types() {
-    /*
-     * Update the name and type on the liquor tab.
-     */
-
-    std::string input_distillery = ui->liquorDistillerInput->currentText().toStdString();
-    std::vector<Drink> selected_drinks = Database::get_beers_by_brewery(storage, input_distillery);
-    std::set<QString> liquor_names;
-    std::set<QString> types;
-
-    // This fixes crashes when changing with rows selected.
-    QSignalBlocker name_input_signal_blocker(ui->liquorNameInput);
-    QSignalBlocker type_input_signal_blocker(ui->liquorTypeInput);
-
-    ui->liquorAbvInput->setValue(0.0);
-    ui->liquorSizeInput->setValue(0);
-    ui->liquorRatingInput->setValue(0);
-    ui->liquorNameInput->clear();
-    ui->liquorTypeInput->clear();
-
-    for (const auto& selected_drink : selected_drinks) {
-        liquor_names.insert(QString::fromStdString(selected_drink.name));
-        types.insert(QString::fromStdString(selected_drink.type));
-    }
-
-    for (const auto& name : liquor_names) {
-        ui->liquorNameInput->addItem(name);
-    }
-
-    for (const auto& liquor_type : types) {
-        ui->liquorTypeInput->addItem(liquor_type);
-    }
-}
-
-void MainWindow::update_wine_names_types() {
-    /*
-     * Update the name and type on the liquor tab.
-     */
-
-    std::string input_winery = ui->wineryInput->currentText().toStdString();
-    std::vector<Drink> selected_drinks = Database::get_beers_by_brewery(storage, input_winery);
-    std::set<QString> wine_names;
-    std::set<QString> types;
-
-    // This fixes crashes when changing with rows selected.
-    QSignalBlocker name_input_signal_blocker(ui->wineNameInput);
-    QSignalBlocker type_input_signal_blocker(ui->wineTypeInput);
-
-    ui->wineAbvInput->setValue(0.0);
-    ui->wineSizeInput->setValue(0);
-    ui->wineRatingInput->setValue(0);
-    ui->wineNameInput->clear();
-    ui->wineTypeInput->clear();
-
-    for (const auto& selected_drink : selected_drinks) {
-        wine_names.insert(QString::fromStdString(selected_drink.name));
-        types.insert(QString::fromStdString(selected_drink.type));
-    }
-
-    for (const auto& name : wine_names) {
-        ui->wineNameInput->addItem(name);
-    }
-
-    for (const auto& wine_type : types) {
-        ui->wineTypeInput->addItem(wine_type);
-    }
-}
-
 void MainWindow::producer_input_changed(const QString&) {
     /*
      *  Change the beer attributes based on the brewery selected in the breweryInput field.
@@ -1247,96 +809,6 @@ void MainWindow::producer_input_changed(const QString&) {
 
     // Update fields based on newly selected beer
     name_input_changed();
-}
-
-void MainWindow::update_beer_types_producers() {
-    /*
-     * Update beer type and brewery when name input changes.
-     */
-
-    // This fixes crashes when changing with rows selected.
-    QSignalBlocker type_input_signal_blocker(ui->beerTypeInput);
-    QSignalBlocker brewery_input_signal_blocker(ui->beerBreweryInput);
-
-    std::string input_beer = ui->beerNameInput->currentText().toStdString();
-    Drink selected_beer = Database::get_drink_by_name(storage, input_beer);
-    std::string beer_type = selected_beer.type;
-    std::string brewery = selected_beer.producer;
-    double abv = selected_beer.abv;
-    double ibu = selected_beer.ibu;
-    int size = selected_beer.size;
-    int rating = selected_beer.rating;
-
-    ui->beerTypeInput->setCurrentText(QString::fromStdString(beer_type));
-    ui->beerBreweryInput->setCurrentText(QString::fromStdString(brewery));
-    ui->beerAbvInput->setValue(abv);
-    ui->beerIbuInput->setValue(ibu);
-    ui->beerSizeInput->setValue(size);
-    ui->beerRatingInput->setValue(rating);
-
-    // Set notes to the notes for beer in the name input
-    std::string notes = get_latest_notes(ui->beerNameInput->currentText().toStdString(), get_current_tab());
-    ui->beerNotesInput->setText(QString::fromStdString(notes));
-}
-
-void MainWindow::update_liquor_types_producers() {
-    /*
-     * Update liquor type and distillery when name input changes.
-     */
-
-    // This fixes crashes when changing with rows selected.
-    QSignalBlocker type_input_signal_blocker(ui->liquorTypeInput);
-    QSignalBlocker brewery_input_signal_blocker(ui->liquorDistillerInput);
-
-    std::string input_liquor = ui->liquorNameInput->currentText().toStdString();
-    Drink selected_liquor = Database::get_drink_by_name(storage, input_liquor);
-    std::string liquor_type = selected_liquor.type;
-    std::string liquor_subtype = selected_liquor.subtype;
-    std::string producer = selected_liquor.producer;
-    double abv = selected_liquor.abv;
-    int size = selected_liquor.size;
-    int rating = selected_liquor.rating;
-
-    ui->liquorTypeInput->setCurrentText(QString::fromStdString(liquor_type));
-    ui->liquorSubtypeInput->setCurrentText(QString::fromStdString(liquor_subtype));
-    ui->liquorDistillerInput->setCurrentText(QString::fromStdString(producer));
-    ui->liquorAbvInput->setValue(abv);
-    ui->liquorSizeInput->setValue(size);
-    ui->liquorRatingInput->setValue(rating);
-
-    // Set notes to the notes for liquor in the name input
-    std::string notes = get_latest_notes(ui->liquorNameInput->currentText().toStdString(), get_current_tab());
-    ui->liquorNotesInput->setText(QString::fromStdString(notes));
-}
-
-void MainWindow::update_wine_types_producers() {
-    /*
-     * Update wine type and winery when name input changes.
-     */
-
-    // This fixes crashes when changing with rows selected.
-    QSignalBlocker type_input_signal_blocker(ui->wineTypeInput);
-    QSignalBlocker brewery_input_signal_blocker(ui->wineryInput);
-
-    std::string input_wine = ui->wineNameInput->currentText().toStdString();
-    Drink selected_wine = Database::get_drink_by_name(storage, input_wine);
-    std::string wine_type = selected_wine.type;
-    std::string wine_subtype = selected_wine.subtype;
-    std::string producer = selected_wine.producer;
-    double abv = selected_wine.abv;
-    int size = selected_wine.size;
-    int rating = selected_wine.rating;
-
-    ui->wineTypeInput->setCurrentText(QString::fromStdString(wine_type));
-    ui->wineSubtypeInput->setCurrentText(QString::fromStdString(wine_subtype));
-    ui->wineryInput->setCurrentText(QString::fromStdString(producer));
-    ui->wineAbvInput->setValue(abv);
-    ui->wineSizeInput->setValue(size);
-    ui->wineRatingInput->setValue(rating);
-
-    // Set notes to the notes for liquor in the name input
-    std::string notes = get_latest_notes(ui->wineNameInput->currentText().toStdString(), get_current_tab());
-    ui->wineNotesInput->setText(QString::fromStdString(notes));
 }
 
 void MainWindow::name_input_changed() {
@@ -1455,53 +927,15 @@ Drink MainWindow::get_drink_attributes_from_fields() {
      */
 
     std::string alcohol_type = get_current_tab();
+
     Drink drink;
 
     if (alcohol_type == "Beer") {
-        drink.drink_year = ui->beerDateInput->date().year();
-        drink.drink_month = ui->beerDateInput->date().month();
-        drink.drink_day = ui->beerDateInput->date().day();
-        drink.name = ui->beerNameInput->currentText().toStdString();
-        drink.type = ui->beerTypeInput->currentText().toStdString();
-        drink.subtype = ui->beerSubtypeInput->currentText().toStdString();
-        drink.producer = ui->beerBreweryInput->currentText().toStdString();
-        drink.vintage = -999;
-        drink.ibu = ui->beerIbuInput->value();
-        drink.abv = ui->beerAbvInput->value();
-        drink.size = ui->beerSizeInput->value();
-        drink.rating = ui->beerRatingInput->value();
-        drink.notes = ui->beerNotesInput->toPlainText().toStdString();
-        drink.alcohol_type = alcohol_type;
+        drink = get_beer_attrs_from_fields(alcohol_type);
     } else if (alcohol_type == "Liquor") {
-        drink.drink_year = ui->liquorDateInput->date().year();
-        drink.drink_month = ui->liquorDateInput->date().month();
-        drink.drink_day = ui->liquorDateInput->date().day();
-        drink.name = ui->liquorNameInput->currentText().toStdString();
-        drink.type = ui->liquorTypeInput->currentText().toStdString();
-        drink.subtype = ui->liquorSubtypeInput->currentText().toStdString();
-        drink.producer = ui->liquorDistillerInput->currentText().toStdString();
-        drink.vintage = -999;
-        drink.abv = ui->liquorAbvInput->value();
-        drink.ibu = -1.0;  // -1 denotes no IBU value
-        drink.size = ui->liquorSizeInput->value();
-        drink.rating = ui->liquorRatingInput->value();
-        drink.notes = ui->liquorNotesInput->toPlainText().toStdString();
-        drink.alcohol_type = alcohol_type;
+        drink = get_liquor_attrs_from_fields(alcohol_type);
     } else if (alcohol_type == "Wine") {
-        drink.drink_year = ui->wineDateInput->date().year();
-        drink.drink_month = ui->wineDateInput->date().month();
-        drink.drink_day = ui->wineDateInput->date().day();
-        drink.name = ui->wineNameInput->currentText().toStdString();
-        drink.type = ui->wineTypeInput->currentText().toStdString();
-        drink.subtype = ui->wineSubtypeInput->currentText().toStdString();
-        drink.producer = ui->wineryInput->currentText().toStdString();
-        drink.vintage = ui->wineVintage->value();
-        drink.abv = ui->wineAbvInput->value();
-        drink.ibu = -1.0;  // -1 denotes no IBU value
-        drink.size = ui->wineSizeInput->value();
-        drink.rating = ui->wineRatingInput->value();
-        drink.notes = ui->wineNotesInput->toPlainText().toStdString();
-        drink.alcohol_type = alcohol_type;
+        drink = get_wine_attrs_from_fields(alcohol_type);
     } else {
         std::cout << "Somehow got illegal alcohol type." << std::endl;
     }
