+++ src/mainwindow.ui
@@ -175,6 +175,11 @@
         <string>Timestamp</string>
        </property>
       </column>
+      <column>
+       <property name="text">
+        <string>Sort</string>
+       </property>
+      </column>
       <item row="0" column="0">
        <property name="text">
         <string>09/07/2020</string>
